requirements.txt
# Web Framework
Flask==3.1.1
Werkzeug==3.1.3

# HTTP Requests
requests==2.32.4
urllib3==2.5.0

# Document Processing
python-docx==1.2.0
PyMuPDF==1.26.3

# HTML Parsing (for training.gov.au scraping)
beautifulsoup4==4.13.4
lxml==6.0.0
soupsieve==2.7

# Data Processing
python-dateutil==2.8.2
openpyxl==3.1.5

# Development & Testing
pytest==8.4.1
pytest-flask==1.3.0

# Environment & Configuration
python-dotenv==1.1.1

# Authentication & OAuth
Flask-Dance==7.0.0
requests-oauthlib==2.0.0

# Optional: Enhanced logging
colorama==0.4.6

# Additional utilities
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2

# JSON Schema Validation
jsonschema==4.25.0 